version 2

+++ controller/admin/product.controller.js
@@ -1,9 +1,13 @@
 const Product = require("../../models/product.model");
+const ProductCategory = require("../../models/product-category.model");
+const Account = require("../../models/account.model");
 const filterHelper = require("../../helpers/filter.helper");
 const paginationHelper = require("../../helpers/pagination.helper");
 const prefixAdmin = require("../../config/system");
+const createTree = require("../../helpers/createTree.helper");
 const { request } = require("express");
 
+// [GET] /admin/products/index
 module.exports.index = async (req, res) => {
     const find = {
         deleted: false
@@ -21,19 +25,42 @@ module.exports.index = async (req, res) => {
     }
     // End Search
     // Pagination
-    
     const countRecords = await Product.countDocuments(find);
     const objectPagination = paginationHelper(req, countRecords);
-    
     // End Pagination
+    const sort ={};
+    if(req.query.sortKey && req.query.sortValue){
+        const sortKey = req.query.sortKey;
+        const sortValue = req.query.sortValue;
+        sort[sortKey] = sortValue;
+    }else{
+        sort["position"] = "desc";
+    } 
+
+    
 
     const products = await Product
     .find(find)
     .limit(objectPagination.limitPage)
     .skip(objectPagination.skipPage)
-    .sort({position: "desc"});
+    .sort(sort);
 
-    res.render("admin/pages/products", {
+    for (const product of products) {
+        const createdBy = await Account.findOne({
+            _id: product.createdBy
+        })
+        const updatedBy = await Account.findOne({
+            _id: product.updatedBy
+        })
+        const deletedBy = await Account.findOne({
+            _id: product.deletedBy
+        })
+        product.createdBy = createdBy?.fullName;
+        product.updatedBy = updatedBy?.fullName;
+        product.deletedBy = deletedBy?.fullName;
+    }
+    
+    res.render("admin/pages/products/index", {
         pageTitle: "Trang danh sách sản phẩm",
         products: products,
         filterStatus: filterStatus,
@@ -42,70 +69,143 @@ module.exports.index = async (req, res) => {
     });
 }
 
+// [PATCH] /admin/products/changeStatus
 module.exports.changeStatus = async (req, res) => {
-    const status = req.params.status;
-    const id = req.params.id;
-    await Product.updateOne({
-        _id: id
-    },{
-        status: status
-    })
-     
-    req.flash("success", "Cập nhật trạng thái thành công !");
-
-    res.redirect(`back`);
-}
-
-module.exports.changeMulti = async (req, res) => {
-    const type = req.body.type;
-    let ids = req.body.ids;
-    ids = ids.split(", ");
-    switch (type) {
-        case "active":
-        case "inactive": 
-            await Product.updateMany({
-                _id: {$in: ids}
-            }, {
-                status: type
-            });
-            break;
-        case "delete-all":
-            await Product.updateMany({
-                _id: {$in: ids}
-            }, {
-                deleted: true
-            })
-            break;
-        case "change-position":
-            for(item of ids){
-                let [id, position] = item.split("-");
-                position = parseInt(position);
-                await Product.updateMany({
-                    _id: id
-                }, {
-                    position: position
-                })
-            }
-            break;
-        default: 
-            break;
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_edit")){
+        res.send("Không có quyền truy cập");
+        return;
     }
+    try{
+        const status = req.params.status;
+        const id = req.params.id;
+        await Product.updateOne({
+            _id: id
+        },{
+            status: status,
+            updatedBy: res.locals.user.id,
+            updatedAt: new Date()
+        })
+        
+        req.flash("success", "Cập nhật trạng thái thành công !");
 
-    req.flash("success", "Cập nhật trạng thái thành công !");
-    res.redirect(`back`);
+        res.redirect(`back`);
+    }catch {
+        req.flash("error", "Cập nhật trạng thái không thành công !");
+
+        res.redirect(`back`);
+    }
 }
 
+// [PATCH] /admin/products/changeMultiStatus
+module.exports.changeMulti = async (req, res) => {
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_edit")){
+        res.send("Không có quyền truy cập");
+        return;
+    }
+    try{
+        const type = req.body.type;
+        let ids = req.body.ids;
+        ids = ids.split(", ");
+        switch (type) {
+            case "active":
+            case "inactive": 
+                await Product.updateMany({
+                    _id: {$in: ids}
+                }, {
+                    status: type,
+                    updatedBy: res.locals.user.id,
+                    updatedAt: new Date()
+                });
+                break;
+            case "delete-all":
+                await Product.updateMany({
+                    _id: {$in: ids}
+                }, {
+                    deleted: true,
+                    updatedBy: res.locals.user.id,
+                    updatedAt: new Date()
+                })
+                break;
+            case "change-position":
+                for(item of ids){
+                    let [id, position] = item.split("-");
+                    position = parseInt(position);
+                    await Product.updateMany({
+                        _id: id
+                    }, {
+                        position: position,
+                        updatedBy: res.locals.user.id,
+                        updatedAt: new Date()
+                    })
+                }
+                break;
+            default: 
+                break;
+        }
+
+        req.flash("success", "Cập nhật trạng thái thành công !");
+        res.redirect(`back`);
+    }catch {
+        req.flash("error", "Cập nhật trạng thái không thành công !");
+        res.redirect(`back`);
+    }
+}
+
+// [DELETE] /admin/products/changeMultI
+module.exports.changeMultiTrash = async (req, res) => {
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_edit")){
+        res.send("Không có quyền truy cập");
+        return;
+    }
+    try{
+        const type = req.body.type;
+        let ids = req.body.ids;
+        ids = ids.split(", ");
+        switch (type) {
+            case "remove-all":
+                await Product.deleteMany({
+                    _id: {$in: ids}
+                })
+                break;
+            default: 
+                break;
+        }   
+        req.flash("success", "Xóa sản phẩm thành công !");
+        res.redirect(`back`);
+    }catch {
+        req.flash("error", "Xóa sản phẩm không thành công !");
+        res.redirect(`back`);
+    }
+}
+
+// [PATCH] /admin/products/deleteItem
 module.exports.deleteItem = async (req, res) => {
-    const id = req.params.id;
-    await Product.updateOne({
-        _id: id
-    }, {
-        deleted: true
-    });
-    req.flash("success", "Xóa sản phẩm thành công !");
-    res.redirect(`back`);
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_delete")){
+        res.send("Không có quyền truy cập");
+        return;
+    }
+    try {
+        const id = req.params.id;
+        await Product.updateOne({
+            _id: id
+        }, {
+            deleted: true,
+            deletedAt: new Date(),
+            deletedBy: res.locals.user.id
+        });
+        req.flash("success", "Xóa sản phẩm thành công !");
+        res.redirect(`back`);
+    } catch {
+        req.flash("error", "Xóa sản phẩm không thành công !");
+        res.redirect(`back`);
+    }
 }
 
+// [GET] /admin/products/trash
 module.exports.trash = async (req, res) => {
     const find = {
         deleted: true
@@ -125,13 +225,22 @@ module.exports.trash = async (req, res) => {
     .find(find)
     .limit(objectPagination.limitPage)
     .skip(objectPagination.skipPage);;
-    res.render("admin/pages/trash", {
+
+    for (const product of products) {
+        const deletedBy = await Account.findOne({
+            _id: product.deletedBy
+        })
+        product.deletedBy = deletedBy?.fullName;
+    }
+
+    res.render("admin/pages/products/trash", {
         products: products,
         filterStatus: filterStatus,
         objectPagination: objectPagination
     });
 }
 
+// [PATCH] /admin/products/recall
 module.exports.recallItem = async (req, res) => {
     const id = req.params.id;
     await Product.updateOne({
@@ -143,7 +252,13 @@ module.exports.recallItem = async (req, res) => {
     res.redirect(`back`);
 }
 
+// [DELETE] /admin/products/delete
 module.exports.removeItem = async (req, res) => {
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_delete")){
+        res.send("Không có quyền truy cập");
+        return;
+    }
     const id = req.params.id;
     await Product.deleteOne({
         _id: id
@@ -153,15 +268,32 @@ module.exports.removeItem = async (req, res) => {
     
 }
 
-module.exports.create = (req, res) => {
-    res.render("admin/pages/create.pug");
+// [GET] /admin/products/create
+module.exports.create = async (req, res) => {
+    const category = await ProductCategory.find({
+        deleted: false
+    });
+    const newRecord = createTree(category);
+    res.render("admin/pages/products/create", {
+        pageTitle: "Trang thêm mới sản phẩm",
+        category: newRecord
+    });
 }
 
-module.exports.createPost = async (req, res) => { 
+// [POST] /admin/products/create
+module.exports.createPost = async (req, res) => {
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_create")){
+        res.send("Không có quyền truy cập");
+        return;
+    }
+   
+    req.body.discountPercentage = 0; 
     if(req.body){
         req.body.price = parseInt(req.body.price);
         req.body.discountPercentage = parseInt(req.body.discountPercentage);
         req.body.stock = parseInt(req.body.stock);
+        req.body.createdBy = res.locals.user.id;
     }
     if(req.body.position){
         req.body.position = parseInt(req.body.position);
@@ -169,44 +301,67 @@ module.exports.createPost = async (req, res) => {
         const countProduct = await Product.countDocuments({});
         req.body.position = countProduct + 1;
     }
-    if(req.file){
-        req.body.thumbnail = `/uploads/${req.file.filename}`;
-    }
+    
+    // if(req.file){
+    //     req.body.thumbnail = `/uploads/${req.file.filename}`;
+    // }
     const record = new Product(req.body);
     await record.save();
     res.redirect(`${prefixAdmin.prefixAdmin}/products`);
 }
 
+// [GET] /admin/products/edit
 module.exports.edit = async (req, res) => {
     const id = req.params.id;
+    const category = await ProductCategory.find({
+        deleted: false
+    });
+    const newRecord = createTree(category);
     const product = await Product.findOne({
         _id: id
     });
-    res.render("admin/pages/edit", {
+    
+    res.render("admin/pages/products/edit", {
         pageTitle: "Trang chỉnh sửa sản phẩm",
-        product: product
+        product: product,
+        category: newRecord
     });
 }
 
+// [PATCH] /admin/products/editPatch
 module.exports.editPatch = async (req, res) => {
-    const id = req.params.id;
-
-    req.body.price = parseInt(req.body.price);
-    req.body.discountPercentage = parseInt(req.body.discountPercentage);
-    req.body.stock = parseInt(req.body.stock);
-    req.body.position = parseInt(req.body.position);
-
-    if(req.file){
-        req.body.thumbnail = `/uploads/${req.file.filename}`;
+    const permissions = res.locals.role.permissions;
+    if(!permissions.includes("products_edit")){
+        res.send("Không có quyền truy cập");
+        return;
     }
-    await Product.updateOne({
-        _id: id,
-        deleted: false
-    }, req.body);
-    req.flash("success", "Cập nhật sản phẩm thành công!");
-    res.redirect(`${prefixAdmin.prefixAdmin}/products`);
+    try{
+        const id = req.params.id;
+
+        req.body.price = parseInt(req.body.price);
+        req.body.discountPercentage = parseInt(req.body.discountPercentage);
+        req.body.stock = parseInt(req.body.stock);
+        req.body.position = parseInt(req.body.position);
+        req.body.updatedBy = res.locals.user.id;
+
+        if(req.file){
+            req.body.thumbnail = `/uploads/${req.file.filename}`; 
+        }
+       
+        await Product.updateOne({
+            _id: id,
+            deleted: false
+        }, req.body);
+        req.flash("success", "Cập nhật sản phẩm thành công!");
+        res.redirect(`${prefixAdmin.prefixAdmin}/products`);
+    }catch {
+        req.flash("error", "Cập nhật sản phẩm không thành công!");
+        res.redirect(`back`);
+    }
+    
 }
 
+// [GET] /admin/products/detail/:id
 module.exports.detail = async (req, res) => {
     const id = req.params.id;
 
@@ -214,7 +369,7 @@ module.exports.detail = async (req, res) => {
         _id: id,
         deleted: false
     })
-    res.render("admin/pages/detail", {
+    res.render("admin/pages/products/detail", {
         pageTitle: "Trang chi tiết sản phẩm",
         product: product
     });
